Update WorkingNFCGUI.ino
Added Comments, several functions, welcomescreen, drawbutton, tapcardscreen, fillscreenblank, checkbuttonpress, onled, offled, startup, iswifiup.

Able to display startup screen, checks if it is connected to wifi if it is not send and error message and light up red led. If it is connected proceed to welcome screen. Clicking one of the plug buttons promts the user to tap their card.